feat(behavior_path_planner): add pull over goal searcher interface (#2069)
