+++ planning/behavior_path_planner/src/scene_module/pull_over/pull_over_module.cpp
@@ -63,15 +63,22 @@ PullOverModule::PullOverModule(
       node, parameters, lane_departure_checker_, occupancy_grid_map_));
   }
   if (parameters_.enable_arc_forward_parking) {
-    const bool is_forward = true;
+    constexpr bool is_forward = true;
     pull_over_planners_.push_back(std::make_shared<GeometricPullOver>(
       node, parameters, getGeometricPullOverParameters(), occupancy_grid_map_, is_forward));
   }
   if (parameters_.enable_arc_backward_parking) {
-    const bool is_forward = false;
+    constexpr bool is_forward = false;
     pull_over_planners_.push_back(std::make_shared<GeometricPullOver>(
       node, parameters, getGeometricPullOverParameters(), occupancy_grid_map_, is_forward));
   }
+
+  // set selected goal searcher
+  // currently there is only one goal_searcher_type
+  const auto & vehicle_footprint =
+    createVehicleFootprint(vehicle_info_util::VehicleInfoUtil(node).getVehicleInfo());
+  goal_searcher_ =
+    std::make_shared<GoalSearcher>(parameters, vehicle_footprint, occupancy_grid_map_);
 
   // for collision check with objects
   vehicle_footprint_ = createVehicleFootprint(vehicle_info_);
@@ -258,59 +265,6 @@ Pose PullOverModule::calcRefinedGoal() const
   return refined_goal_pose;
 }
 
-void PullOverModule::researchGoal()
-{
-  // Find goals in pull over areas.
-  goal_candidates_.clear();
-
-  const auto shoulder_lane_objects =
-    util::filterObjectsByLanelets(*(planner_data_->dynamic_object), status_.pull_over_lanes);
-
-  for (double dx = -parameters_.backward_goal_search_length;
-       dx <= parameters_.forward_goal_search_length; dx += parameters_.goal_search_interval) {
-    // search goal_pose in lateral direction
-    Pose search_pose{};
-    bool found_lateral_no_collision_pose = false;
-    double lateral_offset = 0.0;
-    for (double dy = 0; dy <= parameters_.max_lateral_offset;
-         dy += parameters_.lateral_offset_interval) {
-      lateral_offset = dy;
-      search_pose = calcOffsetPose(refined_goal_pose_, dx, -dy, 0);
-      if (checkCollisionWithPose(search_pose)) {
-        continue;
-      }
-
-      // if finding objects near the search pose,
-      // shift search_pose in lateral direction one more
-      // because collsion may be detected on other path points
-      if (dy > 0) {
-        search_pose = calcOffsetPose(search_pose, 0, -parameters_.lateral_offset_interval, 0);
-      }
-
-      found_lateral_no_collision_pose = true;
-      break;
-    }
-    if (!found_lateral_no_collision_pose) continue;
-
-    constexpr bool filter_inside = true;
-    const auto target_objects = pull_over_utils::filterObjectsByLateralDistance(
-      search_pose, planner_data_->parameters.vehicle_width, shoulder_lane_objects,
-      parameters_.object_recognition_collision_check_margin, filter_inside);
-    if (checkCollisionWithLongitudinalDistance(search_pose, target_objects)) {
-      continue;
-    }
-
-    GoalCandidate goal_candidate;
-    goal_candidate.goal_pose = search_pose;
-    goal_candidate.lateral_offset = lateral_offset;
-    goal_candidate.distance_from_original_goal =
-      std::abs(inverseTransformPose(search_pose, refined_goal_pose_).position.x);
-    goal_candidates_.push_back(goal_candidate);
-  }
-  // Sort with distance from original goal
-  std::sort(goal_candidates_.begin(), goal_candidates_.end());
-}
-
 BT::NodeStatus PullOverModule::updateState()
 {
   // pull_out module will be run when setting new goal, so not need finishing pull_over module.
@@ -346,71 +300,6 @@ bool PullOverModule::isLongEnoughToParkingStart(
   }
 
   return *dist_to_parking_start_pose > current_to_stop_distance;
-}
-
-bool PullOverModule::checkCollisionWithLongitudinalDistance(
-  const Pose & ego_pose, const PredictedObjects & dynamic_objects) const
-{
-  if (parameters_.use_occupancy_grid && parameters_.use_occupancy_grid_for_longitudinal_margin) {
-    constexpr bool check_out_of_range = false;
-    const double offset = std::max(
-      parameters_.longitudinal_margin - parameters_.occupancy_grid_collision_check_margin, 0.0);
-
-    // check forward collision
-    const Pose ego_pose_moved_forward = calcOffsetPose(ego_pose, offset, 0, 0);
-    const Pose forward_pose_grid_coords =
-      global2local(occupancy_grid_map_->getMap(), ego_pose_moved_forward);
-    const auto forward_idx = pose2index(
-      occupancy_grid_map_->getMap(), forward_pose_grid_coords,
-      occupancy_grid_map_->getParam().theta_size);
-    if (occupancy_grid_map_->detectCollision(forward_idx, check_out_of_range)) {
-      return true;
-    }
-
-    // check backward collision
-    const Pose ego_pose_moved_backward = calcOffsetPose(ego_pose, -offset, 0, 0);
-    const Pose backward_pose_grid_coords =
-      global2local(occupancy_grid_map_->getMap(), ego_pose_moved_backward);
-    const auto backward_idx = pose2index(
-      occupancy_grid_map_->getMap(), backward_pose_grid_coords,
-      occupancy_grid_map_->getParam().theta_size);
-    if (occupancy_grid_map_->detectCollision(backward_idx, check_out_of_range)) {
-      return true;
-    }
-  }
-
-  if (parameters_.use_object_recognition) {
-    if (
-      util::calcLongitudinalDistanceFromEgoToObjects(
-        ego_pose, planner_data_->parameters.base_link2front,
-        planner_data_->parameters.base_link2rear,
-        dynamic_objects) < parameters_.longitudinal_margin) {
-      return true;
-    }
-  }
-  return false;
-}
-
-bool PullOverModule::checkCollisionWithPose(const Pose & pose) const
-{
-  if (parameters_.use_occupancy_grid) {
-    const Pose pose_grid_coords = global2local(occupancy_grid_map_->getMap(), pose);
-    const auto idx = pose2index(
-      occupancy_grid_map_->getMap(), pose_grid_coords, occupancy_grid_map_->getParam().theta_size);
-    const bool check_out_of_range = false;
-    if (occupancy_grid_map_->detectCollision(idx, check_out_of_range)) {
-      return true;
-    }
-  }
-
-  if (parameters_.use_object_recognition) {
-    if (util::checkCollisionBetweenFootprintAndObjects(
-          vehicle_footprint_, pose, *(planner_data_->dynamic_object),
-          parameters_.object_recognition_collision_check_margin)) {
-      return true;
-    }
-  }
-  return false;
 }
 
 bool PullOverModule::planWithEfficientPath()
@@ -516,7 +405,10 @@ BehaviorModuleOutput PullOverModule::plan()
 
   } else {  // Replan shift -> arc forward -> arc backward path with each goal candidate.
     // Research goal when enabling research and final path has not been decided
-    if (parameters_.enable_goal_research) researchGoal();
+    if (parameters_.enable_goal_research) {
+      goal_searcher_->setPlannerData(planner_data_);
+      goal_candidates_ = goal_searcher_->search(refined_goal_pose_, status_.pull_over_lanes);
+    }
 
     // plan paths with several goals and planner
     if (parameters_.search_priority == "efficient_path") {
